+++ test/e2e_node/util_machineinfo_unsupported.go
@@ -2,7 +2,7 @@
 // +build !linux
 
 /*
-Copyright 2020 The Kubernetes Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -19,7 +19,42 @@ limitations under the License.
 
 package e2enode
 
+import (
+	"errors"
+
+	"k8s.io/utils/cpuset"
+)
+
 // IsCgroup2UnifiedMode returns whether we are running in cgroup v2 unified mode.
 func IsCgroup2UnifiedMode() bool {
 	return false
+}
+
+func isHTEnabled() bool {
+	return false
+}
+
+func isMultiNUMA() bool {
+	return false
+}
+
+func getSMTLevel() int {
+	return 1
+}
+
+func getUncoreCPUGroupSize() int {
+	return 1
+}
+
+func getCPUSiblingList(cpuRes int64) string {
+	return ""
+}
+
+func getCoreSiblingList(cpuRes int64) string {
+	return ""
+}
+
+// getNumaNodeCPUs retrieves CPUs for each NUMA node.
+func getNumaNodeCPUs() (map[int]cpuset.CPUSet, error) {
+	return nil, errors.New("not implemented")
 }
